[PBE-3809] code clean up
